Mehmet/requirements.txt
openai==0.28
python-dotenv
pandas
tiktoken
fabulous
backoff
evaluate
rouge_score
